unit tests

+++ UbiPod/Models/PodcastEpisode.swift
@@ -5,8 +5,8 @@ struct PodcastEpisode: Decodable, Hashable, Identifiable {
     let shortDescription: String
     let trackName: String
     let id: Int64
-    let releaseDate: Date // "2025-05-26T03:00:00Z"
-    let trackTimeMillis: Int64
+    let releaseDate: Date
+    let trackTimeMillis: Int
 
     enum CodingKeys: String, CodingKey {
         case episodeUrl
